+++ VoidCrewTerminus/ReleaseFiles/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.0.4
+- config setup
+- relic crafting disable patch
 ## 0.0.3
 - workflow fix
 ## 0.0.2